requirements_llava_sam2.txt
# LLaVA + SAM2 + OpenCV Pipeline Dependencies

# Основные зависимости
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.34.0
accelerate>=0.21.0

# OpenCV для постобработки
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Обработка изображений
Pillow>=9.5.0
numpy==1.26.4
scipy>=1.10.0

# SAM2 dependencies
# Устанавливается из GitHub репозитория:
# pip install git+https://github.com/facebookresearch/segment-anything-2.git
# Или для разработки:
# git clone https://github.com/facebookresearch/segment-anything-2.git
# cd segment-anything-2; pip install -e .

# Дополнительные зависимости для SAM2
hydra-core>=1.3.0
iopath>=0.1.9
timm>=0.6.7

# Утилиты
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.31.0

# LLaVA зависимости
protobuf>=3.20.0
sentencepiece>=0.1.99
einops>=0.6.1
gradio>=3.35.0
gradio_client>=0.2.9
markdown2[all]>=2.4.10

# Дополнительные зависимости для стабильности
safetensors>=0.3.1
tokenizers>=0.13.3
huggingface-hub>=0.16.4

# Дополнительные утилиты для видео (для SAM2, если потребуется)
av>=10.0.0

# Для автоматического скачивания моделей
gdown>=4.7.1

# Зависимости для валидации на MVTec AD
kagglehub>=0.2.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0 